Update logging messages and separate dev-tools from the metadata module.
