+++ pyproject.toml
@@ -55,7 +55,7 @@ scicat_ingestor = "scicat_online_ingestor:main"
 scicat_background_ingestor = "scicat_offline_ingestor:main"
 # Dev Tools
 scicat_validate_ingestor_config = "scicat_configuration:validate_config_file"
-scicat_validate_metadata_schema = "scicat_metadata:validate_schema"
+scicat_validate_metadata_schema = "scicat_devtools:validate_schema"
 scicat_synchronize_config = "scicat_configuration:synchronize_config_file"
 scicat_json_to_yaml = "scicat_configuration:json_to_yaml"
     # We don't have to test this json to yaml script because we always check if
